support kotlin 1.9.22
This makes some changes to the client pom to support compiling kotlin

Adds "hotlite.plugins" package to the plugin scanner

Adds K-Example kotlin plugin example.

This is all very basic at the moment but kotlin plugin functionality will eventually be as good as it was on Meteor.

This depends on the recent launcher commit that updates the kotlin std lib for runtime