add client patch peek/poke
With asm and kotlin-reflect added, we can analyze and even modify the client patch at runtime. A simple example has been added that will print out the instructions for client.changeWorld on startup.

This is how we will get our mappings etc so best to get this launcher / client update out of the way now.